Add files via upload
This is a program for the Implementation of a structure-based Academic Information System in CPP. It was assigned to me by Dr. Zahid Halim. 

I and my friend Syed Abdul Rehman Hasan (Manu) wrote the code, there will be some bugs in the code feel free to find them. You can use the codes anywhere. Ty